Extract interface declaration tests
